3DGAN/requirements.txt
dominate
easydict
h5py
numpy
opencv-python
scikit-image
tensorboardX
torchvision
tqdm


